adds helper transform and privatizes several internal objects
