+++ Source/Relativity.Test.Helpers/Relativity.Test.Helpers/Helpers/IHelperTransform.cs
@@ -1,0 +1,9 @@
+﻿namespace Relativity.Test.Helpers
+{
+	interface IHelperTransform
+	{
+		TestAgentHelper AsTestAgentHelper();
+
+		TestEHHelper AsTestEHHelper();
+	}
+}
