+++ Source/Relativity.Test.Helpers/Relativity.Test.Helpers/Helpers/ServiceFactory/URLHelper.cs
@@ -1,0 +1,21 @@
+﻿using System;
+
+namespace Relativity.Test.Helpers.ServiceFactory
+{
+	class URLHelper : API.IUrlHelper
+	{
+		public Uri GetApplicationURL(Guid appGuid)
+		{
+			throw new NotImplementedException();
+		}
+
+		public string GetRelativePathToCustomPages(Guid appGuid)
+		{
+			throw new NotImplementedException();
+		}
+
+
+
+
+	}
+}
