+++ Source/Relativity.Test.Helpers/Relativity.Test.Helpers/Objects/Application/ApplicationInstallException.cs
@@ -2,7 +2,7 @@
 
 namespace Relativity.Test.Helpers.Objects.Application.Exceptions
 {
-	public class ApplicationInstallException : Exception
+	class ApplicationInstallException : Exception
 	{
 		public ApplicationInstallException(string message) : base(message) { }
 	}
